[map-helper-functions] Create additional hooks
