+++ packages/map/src/components/WorldMap/WorldMap.tsx
@@ -14,6 +14,7 @@ import { getGeoFillColor, getGeoStrokeColor, handleMapAriaLabels, titleCase, dis
 import useGeoClickHandler from '../../hooks/useGeoClickHandler'
 import useApplyLegendToRow from '../../hooks/useApplyLegendToRow'
 import useApplyTooltipsToGeo from '../../hooks/useApplyTooltipsToGeo'
+import useGenerateRuntimeData from '../../hooks/useGenerateRuntimeData'
 
 let projection = geoMercator()
 
@@ -21,7 +22,6 @@ const WorldMap = () => {
   // prettier-ignore
   const {
     data,
-    generateRuntimeData,
     hasZoom,
     position,
     setFilteredCountryCode,
@@ -38,6 +38,7 @@ const WorldMap = () => {
   const geoClickHandler = useGeoClickHandler()
   const applyLegendToRow = useApplyLegendToRow(legendMemo, legendSpecialClassLastMemo)
   const applyTooltipsToGeo = useApplyTooltipsToGeo()
+  const generateRuntimeData = useGenerateRuntimeData(state)
 
   useEffect(() => {
     const fetchData = async () => {
@@ -54,9 +55,9 @@ const WorldMap = () => {
 
   // TODO Refactor - state should be set together here to avoid rerenders
   // Resets to original data & zooms out
-  const handleReset = (state, setState, setRuntimeData, generateRuntimeData) => {
-    let reRun = generateRuntimeData(state)
-    setRuntimeData(reRun)
+  const handleReset = (state, setState, setRuntimeData) => {
+    const newRuntimeData = generateRuntimeData(state)
+    setRuntimeData(newRuntimeData)
     setState({
       ...state,
       focusedCountry: false,
@@ -75,7 +76,7 @@ const WorldMap = () => {
   }
 
   // TODO Refactor - state should be set together here to avoid rerenders
-  const handleCircleClick = (country, state, setState, setRuntimeData, generateRuntimeData) => {
+  const handleCircleClick = (country, state, setState, setRuntimeData) => {
     if (!state.general.allowMapZoom) return
     let newRuntimeData = state.data.filter(item => item[state.columns.geo.name] === country[state.columns.geo.name])
     setFilteredCountryCode(newRuntimeData[0].uid)
@@ -187,7 +188,6 @@ const WorldMap = () => {
     // Cities
     geosJsx.push(
       <CityList
-        applyLegendToRow={applyLegendToRow}
         geoClickHandler={geoClickHandler}
         key='cities'
         projection={projection}
@@ -209,9 +209,7 @@ const WorldMap = () => {
           applyLegendToRow={applyLegendToRow}
           displayGeoName={displayGeoName}
           tooltipId={tooltipId}
-          handleCircleClick={country =>
-            handleCircleClick(country, state, setState, setRuntimeData, generateRuntimeData)
-          }
+          handleCircleClick={country => handleCircleClick(country, state, setState, setRuntimeData)}
         />
       )
     }
@@ -223,12 +221,7 @@ const WorldMap = () => {
     <ErrorBoundary component='WorldMap'>
       {hasZoom ? (
         <svg viewBox='0 0 880 500' role='img' aria-label={handleMapAriaLabels(state)}>
-          <rect
-            height={500}
-            width={880}
-            onClick={() => handleReset(state, setState, setRuntimeData, generateRuntimeData)}
-            fill='white'
-          />
+          <rect height={500} width={880} onClick={() => handleReset(state, setState, setRuntimeData)} fill='white' />
           <ZoomableGroup
             zoom={position.zoom}
             center={position.coordinates}
@@ -260,8 +253,6 @@ const WorldMap = () => {
         (state.general.type === 'world-geocode' && hasZoom) ||
         (state.general.type === 'bubble' && hasZoom)) && (
         <ZoomControls
-          // prettier-ignore
-          generateRuntimeData={generateRuntimeData}
           handleZoomIn={handleZoomIn}
           handleZoomOut={handleZoomOut}
           position={position}
